// src/components/UI/CarLine/enum.ts
export enum CarLineClassNames {
  WRAPPER = 'car-line',
  CAR_LINE_HEADER = 'car-line-header',
  SELECT_BTN = 'select-car',
  REMOVE_BTN = 'remove-car',
  CAR_NAME = 'car-name',
  CAR_LINE_BODY = 'car-line-body',
  CAR_LINE_BODY_BTN = 'car-line-body-btn',
  START_WRAPPER = 'start',
  CAR_SVG_WRAPPER = 'car-svg',
  FINISH_SVG_WRAPPER = 'finish-svg',
  GRAY = 'gray',
}

export enum CarLineButtonsText {
  SELECT = 'Select',
  REMOVE = 'Remove',
  START = 'A',
  STOP = 'B',
}